Put DeserializeStream into separate file
